Fix Comparison Warning

+++ Source/Core/Renderer/Renderer.cpp
@@ -142,7 +142,7 @@ void Renderer::UpdateLoop() {
     // render the loaded models
     std::vector<ERS_OBJECT_MODEL*> LoadedModelPointers = AssetLoader_->GetModelsToDraw();
 
-    for (long i; i < size(LoadedModelPointers); i++) { 
+    for (long i; i < (long)size(LoadedModelPointers); i++) {
